+++ docs/changelog.rst
@@ -35,6 +35,11 @@ which comes with several pre-built functions (named methods) and attributes. See
 
 Please see the documentation of :py:class:`yasa.Hypnogram` for more details.
 
+.. important::
+  The adoption of object-oriented :py:class:`yasa.Hypnogram` usage brings along critical changes to several YASA function, for example:
+
+  * :py:func:`yasa.simulate_hypno` now returns a `yasa.Hypnogram` instead of a :py:class:`numpy.ndarray`.
+
 ----------------------------------------------------------------------------------------
 
 v0.6.3 (December 2022)